Swap find for where to grab category names
Using `find` throws an error on an item's edit history page when the category has been deleted. Updates the query to use `where` instead. If the number of categories found is less than the number of keys, add the string `deleted category` to signify that there was also a now missing value associated with the item. Adds a test to exercise this behavior.

Issue #41

ruby -I ./test test/system/items_test.rb output:

Finished in 10.779663s, 0.6494 runs/s, 1.4843 assertions/s.
7 runs, 16 assertions, 0 failures, 0 errors, 1 skips