Add member_renewable flag to borrow_policy
Added a member_renewable flag to the borrow_policy that allows admins to
configure policies that will enable members to renew a loan themselves.
Refactored tests to move logic for checking if a loan is member
renewable to the model. Added a button to the member loans list that
allows a member to renew their loans if the loan is within the original
loan duration, has not exceeded the maximum number of renewals according
to the borrow policy, and the policy allows members to renew the loan.