Allow dropping the DB in Docker
The previous approach required the web container to already be started in order to run DB migrations. Since the application was already connected, Postgres would refuse to drop the database until the application was shut down, being a chicken and egg situation. With this dependency specified, the database is started up to run actions against it.