Extract method for item hold count
We need to know how many active holds a member has on a item. This is
important when knowing if member already has a hold so we can disable
buttons to add a hold, and when the item allows multiple holders for
the same member showing how many holds the currnelty have. Extracting
this out of the controller and into the hold model to make this
reusable.