Implement place on hold button functionality for items (#223)
* Implement place on hold button functionality for items

* Use devise method instread of current user?

* Add custom validation for hold creation

* WIP: Start adding some tests

* Add test for borrow policy single or multiple holds

Some items allow a member to have multiple active holds on the same
item. Most items only allow a member to one acitve hold on the item at
a time. This is dictaged by the borrow policy for the item.

* Extract method for item hold count

We need to know how many active holds a member has on a item. This is
important when knowing if member already has a hold so we can disable
buttons to add a hold, and when the item allows multiple holders for
the same member showing how many holds the currnelty have. Extracting
this out of the controller and into the hold model to make this
reusable.

* Fix logic for validation on hold model

* Add links for login and signup to message on item show

When a user is not singed in, but is viewing a item, we want them to
know they can place a hold if they sign in or sign up for a
membership. Adding links to make it eaiser for the user to take the
action.

* Refactor the member holds contorller

* Fix sytle violations

* Add system test for place a hold as a member

* Use uniquely_numbered to determine of multiple holds can be placed

* Remove validaiton on holds

* Fix member holds test fixture setup

Co-authored-by: Terry Finn <tfinn@hey.com>