Set role for admin user seed to admin
After pulling latest development and attempting to login with the
admin@chicagotoollibrary.org user, I was unable to view the admin
dashboard. After checking in my console, I saw that this user did
not have the role of admin. This PR updates the seed to set its
role to admin.

(I could definitely be missing something, but this is the user I
had previously been logging in as when testing admin functionality!)

system tests:
58 runs, 278 assertions, 0 failures, 0 errors, 4 skips

unit tests:
208 runs, 624 assertions, 0 failures, 0 errors, 2 skips