Update setup script to run yarn
Running `bin/setup` on a new project was failing with the following
error:

```
error Couldn't find a binary named "mjml"
rails aborted!
Couldn't find the MJML 4. binary. Have you run $ yarn add mjml?
```

This was failing when attempting to create the database. I believe this
is related to bootsnap loading the environment and the `mjml-rails` app
expecting the `mjml` binary to exist.

By running `yarn` first the `mjml` npm package will be installed and the
binary will be available, enabling `bin/setup` to complete
successfully.