allow admins to cancel holds from an appointment
In an effort to allow admins to cancel the hold for an item at the same
time as they remove an item from an appointment, a button was added to
the admin appointment show page.

We opted to add another parameter to the existing
remove_appointment_hold controller action and base the canceling of a
hold on that. We chose to do this instead of making a new action to
perform both destroys.

* Turned the "remove item" link into a button.
* Made the "Cancel Hold" action a button.
* Moved these buttons to appear below the existing "Check-out" button

Co-authored-by: Valerie Foster <fosterv2@outlook.com>
Signed-off-by: Robb Kidd <robb@thekidds.org>