Bump webpacker from 4.0.2 to 5.4.0
In versions of webpacker prior to 5.4.0, webpacker has a dependency on
`node-sass`, which has a dependency on `node-gyp`. This dependency seems
to be causing problems with Apple's slicon chip. In 5.4.0, this
dependency is removed in favor of `sass`, which is the default
implemenation written in dart. `sass` does not depend on `node-gyp`.