Add method to ignore js errors when they are intentional
This adds a method that clears the js console errors and prevents them from littering the test output when they are caused intentionally.

Any noise in the test output has the potential to distract a developer from seeing actual errors versus the extraneous output. In cases where you will cause a console error intentionally, you should be able to prevent that from showing in the test output.

There are two ways you can use the `ignore_js_errors` method in your tests: inline and block form.

In the block form, surround the code you know will cause the errors.

```ruby
test "it gives an error when you submit without filling in the form" do
  ignore_js_errors(reason: "Creating without filling in the fields gives a 422") do
    visit edit_item_url(@item)
    click_on "Submit"
    expect(page).to have_text("field needs to be filled out")
  end
end
```

The same test can be written the inline form of `ignore_js_errors` by calling it at the end of the test.

```ruby
test "it gives an error when you submit without filling in the form" do
  visit edit_item_url(@item)
  click_on "Submit"
  expect(page).to have_text("field needs to be filled out")
  ignore_js_errors(reason: "Creating without filling in the fields gives a 422")
end
```