[#672] Adds availability filter to public inventory index (#791)
* Adds availability filter to public inventory index

The `current_page_params` helper has been added to retain the existing query params. For example, this will support the following use case: the user initially clicks on the button to sort by "item number", and then they click on the button to filter by "Available now", then the public inventory will include only items that are available, and it will sort the items by item number.

* Filtering by availability, while viewing category

When the user clicks on a category link, they can also filter that particular category by items that are "Available now' or "All items".

* Improves UI of filtering and sorting links

* Adds test coverage for viewing active/all items

* Excludes items w/holds when filtering by available

When the user is only viewing available items, the results no longer include items with holds. The goal is to only show tools that have the active status and are neither checked out or have holds waiting.

* Resolves flaky spec

Prior to this commit, this spec would sometimes fail, due to "random" ordering changes.

* Handles case w/available filter, and with category

When the user has the "available" filter selected, and they are viewing a category of items, any items with a hold should *not* be included in the results.

* Refactors Item.without_holds scope

This scope no longer uses the counter cache.

* Item.with_uniquely_numbered_borrow_policy scope

* Item.without_any_holds scope

* Displays only available items

* Matches terminology used in other parts of app

* Adds test coverage for uncounted items

* Excludes loaned items when viewing only available