Move neon updating into Neon::Client class.
In order to update and not re-create addresses on the account,
the parameters need to be updated in a way that requires access
to the account payload retrieved from the server. Moving the
updating logic into the Client makes it easier to share this state
between different parts of the operation.