remove vestigial category record from test
All of the assertions in this test are scoped to items in a specific category. The presence of a second unnamed category is not required to test the scoping and its presence in the spec could confuse future readers. (It confused me at least!)

I meant to include this removal in PR #813, but got excited with merging