Clarify info on hosts file & add urls to visit
This clarifies the multi-tenancy section on how to edit the hosts file
to add two libraries and point both to localhost.

More importantly, it adds a short sentence with links to access these
two libraries locally.

It wasn't clear that the new urls still needed the port to work. With
direct links that include the port, new contributors that might not know
it was needed won't have problems accessing the two libraries.