Remove Loan#within_borrow_policy_duration method.
This method had a reference to Time.current that was causing tests to break when loans straddled a daylight savings time change and was additionally really difficult to reason about. There was a comment mentioning that it could be removed, so I went ahead and did that.

It's possible that this will make renewals more open than we wish, but we can lock things back down when we identify the situation where that is relevant.