configure Pagy to return last page on overflow
By default, Pagy throws an exception when given a page parameter greater
than the number of pages available in a result. This change adds in
Pagy's Overflow extra[1] (included in the gem already) and configures
overflow to return the last page of results when page param is too large
(in other words, when page overflows).

[1] https://ddnexus.github.io/pagy/extras/overflow