Only list overdue items in overdue notice emails
The filtering is done higher up in the chain so that we don't need logic
in the views. The frontend only receives the loan summaries that need to
be displayed. This approach is more testable and more maintainable.

Card: https://github.com/rubyforgood/circulate/projects/4#card-70957698
Issue: https://github.com/rubyforgood/circulate/issues/814