Replace webpacker with esbuild (#1128)
# What it does

This replaces the use of webpacker for bundling JavaScript and CSS with
a combination of esbuild for JS and DartSass for CSS.

# Why it is important

* The current webpacker config means that rebuilding the JS is pretty
slow.
* Rails has deprecated webpacker and future energy is directed in other
directions. This change aligns the project with upstream efforts.

# Implementation notes

* We were previously using browserlist's `defaults` to translate new JS
features into backward-compatible forms. With `esbuild`, the options are
more limited. Using `--target=chrome58,firefox57,safari11` to restrict
what JS features are used should be OK for now and is more conservative
than what we were shipping before.

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.
- [ ] I am interested in feedback but don't need to make the changes
myself.
- [ ] I don't have time or interest in making additional changes to this
work.
- [ ] Other or not sure (please describe):