Fix permissions issue and missing assets in dockerfile (#1135)
# What it does

This fixes a permissions issue I was experiencing on WSL2 + Docker. It
also fixes missing assets by baking them into the image.

# Why it is important

Improves on boarding

# Implementation notes

Assuming docker-compose is being used in a dev environment, there's
another service that should be running for webpack with the dev reload
enabled. I left this out as we're moving to esbuild anyways. I can
implement that service once that PR is merged.

# Your bandwidth for additional changes to this PR

* I have the time and interest to make additional changes to this PR
based on feedback.