Configure production logs to go to appsignal. (#1143)
# What it does

* Configures the production logs to log in a single-line format to
appsignal

# Why it is important

This will make our logs more searchable and also adds a `user_id`
property so we can see what actions a given user takes. A few times now
we've needed to understand a bug report from a user and didn't have this
as a diagnostic tool.

# Implementation notes

* Added the `lograge` gem.
* In production, configured logs to go to appsignal using the logfmt
format.
* The development logs and normal Rails logs that go to Heroku should
remain unchanged.=

# Your bandwidth for additional changes to this PR

- [x] I have the time and interest to make additional changes to this PR
based on feedback.