Add tests to expose bug in appointment mail sending.
When a member schedules a new appointment with the same time as
an existing one, we merge the appointments together. However, we were still
sending an email with an appointment argument that was the appointment that
no longer existed after the merge. This led to deserialization errors when the
mail job was later processed.