Add Nix development setup
This setup works well for me, and should allow anybody familiar with Nix
to get up and running with the project really quickly.

I documented this as an advanced development setup in the same vein as
Docker, leaving the main README instructions as the happy path.

At any point this can be safely ripped out by reverting this commit or
by manually doing the following:

 1. remove .envrc, flake.nix, flake.lock, and NIX.md
 2. clear the the extra entry in .gitignore.
 3. remove the extra configurable in application_system_test_case.rb