Add Nix development setup (#1156)
I'm loathe to be the guy who comes in and messes with the development
environment as the first thing he does, but here I am being that guy. 😅

I've been pretty into Nix for my development environments for the past
couple of years for its ability to lock the full set of dev dependencies
in a consistent way across machines and operating systems.

I'll use this setup myself either way, but it's a bit less overhead for
me if I'm able to check in the configs vs dancing around local
gitignores to keep it working. Plus it gives other people the
opportunity to use it!

I tried to document everything nicely, keep it off to the side, and make
it minimally invasive. But I totally understand if you'd rather keep
this out of the repo! Just let me know. 🙇🏻

----

Commit message:

> This setup works well for me, and should allow anybody familiar with
Nix to get up and running with the project really quickly.
> 
> I documented this as an advanced development setup in the same vein as
Docker, leaving the main README instructions as the happy path.
> 
> At any point this can be safely ripped out by reverting this commit or
by manually doing the following:
> 
>  1. remove .envrc, flake.nix, flake.lock, and NIX.md
>  2. clear the the extra entry in .gitignore.
>  3. remove the extra configurable in application_system_test_case.rb