Add note about adding library email to contacts to welcome messages (#1157)
# What it does

Resolves #1145 

# Why it is important

This could help with email deliverability!

# UI Change Screenshot

![2023-10-14 at 12 49
05@2x](https://github.com/chicago-tool-library/circulate/assets/37534/2d53ac31-81ef-4bba-b704-f8eb507f981c)

# Implementation notes

- Caught some duplication in the text version of the welcome email -
added a covering test and fixed it.
- Some callers of the `instructions` partial use locals to pass
`library` and others use instance variables. I think ideally moving to
explicit locals is a bit cleaner, but I decided to continue to handle
both in the partial for now rather than taking on the larger task of
changing all the callers.
- Added basic unit tests for a couple other member emails which caught a
deprecation warning on rendering a partial with a filename suffix.