Configure Solargraph with Standard Ruby (#1168)
# What it does

The current checked-in configuration runs Solargraph with a default
Rubocop config which goes a bit wild with suggestions. This should make
the LSP only suggest things that match the rules from Standard Ruby.

# Why it is important

Having the LSP config match the linter makes it a bit faster to write
complaint code since you'll get suggestions in-line when using an
LSP-supporting editor.