Update render and redirect calls as required by Turbo.
Turbo requires that redirects returned by form submissions have a status of
303 (see other) so that the browser will follow them with a GET request and
not use the HTTP method of the submission request. This is semantically
correct, but a change in behavior from how Rails UJS handled things.

Additionally, Turbo will not render anything when a 200 response is returned
by a form submission, instead requiring that a 422 (unprocessable entity)
status is returned in order for it to render the new HTML and its validation
error messages. This is going to become the Rails default in the future.