Add controller tests for renewals/payments (#1209)
# What it does

Adds a controller test for `Renewals::PaymentsController`, based heavily
on the one we have for `Signup::PaymentController`. This is a followup
to #1192, which fixed a bug but did not include tests for timing
reasons.

# Why it is important

This is a critical flow in the application, and we'd like to ensure that
it is still working as intended 😄

# Implementation notes

* There's a decent amount of duplication between
`Renewals::PaymentsController` and `Signup::PaymentController`, and the
same is true of their tests. It is looking like we may end up re-working
the payment flows as a part of upcoming work, so I opted to not do
anything too invasive at this time.

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.
- [ ] I am interested in feedback but don't need to make the changes
myself.
- [ ] I don't have time or interest in making additional changes to this
work.
- [ ] Other or not sure (please describe):