Add member with appointment to devdata (#1211)
# What it does

This PR renames the devdata members so that their emails are more
representative (e.g. `member_with_holds@example.com` instead of
`member1@example.com`). It also adds a new member to that dataset with
an upcoming appointment.

# Why it is important

I'm hopeful this will make it easier to know what devdata member has
which attributes, and that having one with an appointment will reduce
how much work someone has to do if they are going to work on something
related to appointments.

# Implementation notes

* I also tweaked the script to add pronouns and a member number to these
members so that they were more similar to our real members.

# How to test

* Run `bin/reset`, then login with `member_with_appointment@example.com`
and verify that they have an appointment and that you can see it from
the admin side as well.

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.
- [ ] I am interested in feedback but don't need to make the changes
myself.
- [ ] I don't have time or interest in making additional changes to this
work.
- [ ] Other or not sure (please describe):