Remove special case for numbering A tools. Fixes #1223. (#1224)
# What it does

We originally treated A tools (those that we don't individually number)
specially, with the idea being that having their numbers be 3 digits
would be another signal that they were A tools. This worked, but...
we're gonna run out of numbers since B tools start at 1000.

So the solution is to remove that special case for A tools. Once this is
merged, all new items will get the next available number in the system.

# Why it is important

I don't know what will happen if we hit 1000 for A tools, but let's not
find out.

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.
- [ ] I am interested in feedback but don't need to make the changes
myself.
- [ ] I don't have time or interest in making additional changes to this
work.
- [ ] Other or not sure (please describe):