Improve devdata somewhat (#1233)
# What it does

This improves a few things about how we do devdata based on a discussion
at tonight's meetup:

# Why it is important

It's important that we reduce the number of sharp edges in the
development environment so that we're as welcoming to new volunteers as
we can be.

# UI Change Screenshot

![Screenshot 2024-01-08 at 9 23
56 PM](https://github.com/chicago-tool-library/circulate/assets/3331/6f794f38-74c2-4606-a482-9334d52c4547)

# Implementation notes

* We now generate a lot of events, (two a day, three days a week, each
week for the next year). This means that folks won't have to run
`bin/reset` if it's been more than a week since they last ran it. I
still think it's a good idea to run it fairly often, but what we have
now is a pretty bad experience considering that contributors are likely
to go more than a week while working on a feature.
* Hardcoding an encrypted password should mean that you don't get logged
out when you run `bin/reset` anymore!
* I updated the Documents to contain human-friendly default values so
that it's more obvious when they appear in the app. Before, for example,
when signing up as a new user you'd just see an empty page because there
wasn't any content set in the Documents.

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.
- [ ] I am interested in feedback but don't need to make the changes
myself.
- [ ] I don't have time or interest in making additional changes to this
work.
- [ ] Other or not sure (please describe):