Replace Turbolinks with Turbo (and upgrade Stimulus) (#1181)
# What it does

* Replaces Turbolinks in the app with the newer version of the library,
Turbo.
* Upgrades Stimulus to the latest version
* Replaces our PortalRendering code with equivalent Turbo-based code.

Fixes #1184.

# Why it is important

Turbolinks is EOLed, and we need to upgrade to Turbo to get back on a
supported library. And by using Turbo instead of custom code, we have
less to maintain and it's easier for contributors to understand how
things work as there are many resources out there for learning Turbo.

# Implementation notes

* This is based on the main parts of #810, without the use of real-time
Turbo Streams as that requires Redis. We can add that later if needed,
but I don't want to pull in that dependency as a part of this.

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.
- [ ] I am interested in feedback but don't need to make the changes
myself.
- [ ] I don't have time or interest in making additional changes to this
work.
- [ ] Other or not sure (please describe):

---------

Co-authored-by: derrick lannaman <derrick.lannaman@powerhrg.com>