Fix "remove item" button in appointment (#1235)
# What it does

Fixes #989. Previously, the "Remove Item" button removed the item from
the appointment *and* canceled the member's hold on the item. Now the
"Remove Item" button works as expected: it removes the item from the
current appointment but does not remove the member's hold on the item.

# Why it is important

Fixes #989

# UI Change Screenshot

No UI changes.

# Implementation notes

* The `cancel_hold` parameter was being passed to the
`remove_appointment_hold` method as a string rather than a boolean
value, which caused the conditional statement to always evaluate as
true. I fixed the issue by coercing to a boolean value, which I think
should work whether the parameter is passed in as a string or a boolean.
Let me know if it would be better to fix it upstream, i.e., by making
sure we're passing a boolean into the function rather than a string.
* I did not create or update any unit tests. Let me know if we should
add a unit test for removing an item (I would likely need some help with
this).

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.
- [ ] I am interested in feedback but don't need to make the changes
myself.
- [ ] I don't have time or interest in making additional changes to this
work.
- [ ] Other or not sure (please describe):