Setup Appsignal to track front end errors. (#1244)
# What it does

We're using [AppSignal](https://www.appsignal.com) to track server-side
exceptions already. This PR sets things up so that we are notified if
there are client-side errors in the app's JavaScript.

With the recent move to the latest Turbo, it seemed like a good time to
get this in place in case there are any issues in the new front end
code.

# Why it is important

We want to know if the JS code breaks in anyone's browser.

# Implementation notes

* This will only track errors that occur within Stimulus controllers.
That is probably a good place to start.
* AppSignal recommends **not** catching all client-side errors (which is
why their library doesn't by default) due to the amount of noise
generated by browser extensions, ad blockers, etc.
* It wouldn't be a bad idea to consider moving our rebuild config into
config files as we're passing a lot of options. I'll leave this for a
future PR, though.
* I set the `APPSIGNAL_FRONTEND_KEY` environment variable on both
production and staging.

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.
- [ ] I am interested in feedback but don't need to make the changes
myself.
- [ ] I don't have time or interest in making additional changes to this
work.
- [ ] Other or not sure (please describe):