Get a fresh schema dump post Rails upgrades
This gets us a fresh dump of the schema by running `db:drop`,
`db:create`, and `db:migrate`.

The most important diff here is the `precision: nil` additions, which
you can read about at length here:
https://github.com/rails/rails/issues/44571

Because of this change, we can drop the special timestamp handling we
had added in the interim upgrades.

Also we're telling the standardrb linter not to touch the schema.rb file
since it's generated and we don't want the linter to fight it.