Improve readability of member appointment page (#1262)
# What it does

Fixes #1236. Readability improvements to the member appointment page. It
fixes the points called out in the issue, plus a few other small
changes.

* Decrease the font size of the date and time
* Move images to the left of descriptions
* Remove horizontal lines between items
* Add vertical margin between the "Appointments" section and the "How do
appointments work" section when viewed on small screens (e.g., phones).
There was already a horizontal margin when viewed on a wide screen.

This PR also adds images to all items in the development version of the
app. The same image is used for all items.

# Why it is important

#1236 

# UI Change Screenshot

Before:


![image](https://github.com/chicago-tool-library/circulate/assets/38900370/ab3436f5-3972-4129-ba5b-7f8e6fd632b8)

After:


![image](https://github.com/chicago-tool-library/circulate/assets/38900370/8de36b1a-c5bc-41fa-a0a4-3b59f033fe56)

# Implementation notes

The main issue was a div tag that wasn't closed properly. That caused
the date/time header to flow directly into the card body rather than
starting a new line.

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.
- [ ] I am interested in feedback but don't need to make the changes
myself.
- [ ] I don't have time or interest in making additional changes to this
work.
- [ ] Other or not sure (please describe):