Initial support for SMS notifications (feature flagged) (#1197)
# What it does

Adds SMS notification support with overdue notices as a first reminder.

# Why it is important

Hopefully SMS notifications will help make it more likely that members
will remember their appointments and return overdue items.

# UI Change Screenshot

![2024-01-22 at 16 21
13@2x](https://github.com/chicago-tool-library/circulate/assets/37534/85d4b02c-36ba-4582-bce6-b05e1f90178e)

# Implementation notes

* Overall the design is meant to mirror ActionMailers. I started by
using a gem called `textris` which takes this shape, but the gem itself
had a lot of compatibility issues so I ended up deciding to roll our own
simple version of the design.
* The unit tests stub out the Twilio client, but interacting with this
to send real messages requires Twilio credentials. I'll document more
about the Twilio integration in the README in an upcoming PR.
* SMS Notifications are behind a feature flag `FEATURE_SMS_REMINDERS`
which is set to on in development but will be off in production until we
are happy with testing in staging.
* For saving `Notification` records, I decided to use the same model and
loosen the schema to be able to cover SMS and emails.
* Changed the user factories/seeds to default to having a phone number
that will generate a successful SMS response when using Twilio test
credentials.
* The UI changes are not behind the feature flag, but I feel like
they're fine to leave in as we're ok to collect user preferences ahead
of the feature being on. I can change this if folks feel differently
though.
* There's a small nix devenv tweak piggybacking on here that fixes a
warning when `convert` is called in the tests.